requirements.txt
amqp==5.3.1
APScheduler==3.11.0
asgiref==3.9.1
attrs==25.3.0
billiard==4.2.1
celery==5.5.3
cffi==1.17.1
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coverage==7.10.2
cryptography==45.0.5
Django==5.2.4
django-cors-headers==4.7.0
django-debug-toolbar==6.0.0
django-environ==0.12.0
django-extensions==4.1
django-filter==25.1
django-rq==3.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
factory_boy==3.3.3
Faker==37.5.3
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.6
graphql-relay==3.2.0
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
numpy==2.3.2
packaging==25.0
pandas==2.3.1
pluggy==1.6.0
promise==2.3
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pycparser==2.22
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.1
pytest-cov==6.2.1
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
referencing==0.36.2
rpds-py==0.26.0
rq==2.4.1
scipy==1.16.1
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.2.13
